requirements.txt
colorama>=0.4.6
cryptography>=43.0.0
ExifRead>=3.0.0
json5>=0.9.14
Pillow>=10.4.0
piexif>=1.1.3
urllib3>=2.0.4
python-dateutil>=2.9.0.post0
ffmpeg>=1.4
ffmpeg-python>=0.2.0
